Remove entries from need final +2 that have a -1
A -1 on a commit means that more discussion needs to happen and it's
not ready for merge. However, the dashboard view for needs final +2
wasn't checking this. The perfect example was tempest change
I7c3cf17bfd8d273823774cd162ae011161c36d3d
which I feel shouldn't be in that list. This commit fixes this by
excluding any commit that has a -1 or -2.